Completed Task#178 under US#129
I have created a new JavaScript file for our Timer. This will take the
place of the previous views.py function that I had dedicated to the timer.
For now it has simply been incorporated into the begin.html file and will
be incorporated into the rest of the UI in further tasks. As of right now
when the "Start Trauma" button is pressed the timer is created and the time
is printed to the browsers console.

I have also removed all remnants of the old timer implementation.