Complete Task #187 under US#129
The time elapsed that is displayed to the user is now persistent. This
means that if the user starts a session, navigates away from the
decision page, and then comes back the time elapsed will still be
accurate based on when the session was initially started. This also
factors in how long the user spent away from the page.

I also added comments to all my new code and renamed the js file.