Completed task #183 under US#129
I have implemented a javascript file that will be the basis for the
general alert system. Within this script the system checks every second
to see if an alert must be thrown based on certain parameters. I also
added some comments to my code in hope of clarify what I have done.