Made small logic modification
I made a small change to the logic in the general alert system. Now,
rather than automatically setting the alerts to be checked every second,
the system will first check if the alert has already been dismissed. If,
it has there is no need to run those time/CPU consuming tasks of checking
for new alerts every second. This should improve the apps speed and
performance later down the road.