Complete US#203 (Tasks #216 ans 217)
THe GCS value is now calculated based on user input. This calculation is done once the use enters a value for each of the thre GCS areas (motor, verbal, and eye). The value is then stored so it can be referenced later on. If the value is less than 13 and the ETCO2 is 40 - 50 an appropriate alert is now thrown. The GCS value will be checked every second while the app is running so that it can be displayed and used for other evaluations.