Completed US#201 (Tasks #220, #221, #222)
The application now calculates the shock level based on user input and throws the correlated alert when applicable. It also reads the HR and BP based
on user input and throws the correlating alerts. Based ont he documentation from Dr. Notrica, the BP alerts need the childs age as well. Currently, there is no where for the user to input this so I simply threw the alert without factoring that in.