Completed US#200 Task#237
I have added a new js file named checkVitals.js that will be responsible for constantly looping and checking if the vitals on the summary page need to be updated. Heart rate, blood pressure, and shock level are all updating accordingly and are displaying on the summary page. Next step is to simply neaten up the vitals display.