Completed US#242 Task#243
The general ETT alerts has now been implemented. When the user clicks "initiated" next to ETT in the Airway protocol the system check to see if the etco2 has been recorded. If it has not, the alert is thrown. As soon as the etco2 level is recorded, the alert is dismissed.