Complete Task#308 under US#307
I have added ETC02 and GCS to the vitals display on the summary page and moved them from local storage into the database. They now working using both Ajax and local storage to cut down on the number of requests we are sending to our server at a given time.