Fixed race condition issue with vitals
I had to add async/await functions to each of the functions responsible for reading the vitals. This was to ensure that the value was actually read before the function continued. I have set a timeout on reading the values to compensate for the user typing. in other words, this all prevents the value of "123" from being read as "1" "2" nad "3".