Created view metrics functionality
I have added the 'view metrics' page and begun to take data from the database and display it on the page so the user can print it as a pdf. I also modified the database so that, for each primary survey step, the entire history of values is save with a correlating timestamp. I will work on formatting the view metrics page next